Refactor gridding module
- Add new grid_vis_loop_data method that allows to easily grid
  data coming from the vis_loop
- Move some methods to new gridding.utils submodule
- Update gridding script
