+++ pyvisgen/gridding/__init__.py
@@ -1,13 +1,11 @@
 from .gridder import (
-    calc_truth_fft,
     convert_amp_phase,
     convert_real_imag,
-    create_gridded_data_set,
     ducc0_gridding,
     grid_data,
-    open_data,
-    save_fft_pair,
+    grid_vis_loop_data,
 )
+from .utils import calc_truth_fft, create_gridded_data_set, open_data, save_fft_pair
 
 __all__ = [
     "calc_truth_fft",
@@ -16,6 +14,7 @@ __all__ = [
     "create_gridded_data_set",
     "ducc0_gridding",
     "grid_data",
+    "grid_vis_loop_data",
     "open_data",
     "save_fft_pair",
 ]
